Refactoring on  GeographyModel phase #1
